why I did not commit this, I have no idea

+++ lib/faster_gem_scripts.rb
@@ -1,12 +1,25 @@
-require 'sane'
 
 class FasterGemScripts
   def self.overwrite full_path
-   contents_should_be = File.read __dir__ + '/template'
-   File.write(full_path, contents_should_be)
-   bin_loc = full_path + '_bin_location'
-   # delete old
-   File.delete(bin_loc) if File.exist?(bin_loc)
+   contents_should_be = File.read File.dirname(__FILE__) + '/template'
+   if File.file?(full_path)
+     puts 'faster_gem_scripts: optimizing ' + full_path + '...'
+     File.open(full_path, 'wb') do |f| f.write(contents_should_be); end
+     bin_loc = full_path + '_bin_location'
+     File.delete(bin_loc) if File.exist?(bin_loc)
+   else
+    puts "unable to find the binary:" + full_path + " you may want to find it and overwrite it yourself using C:\>faster_gem_script file_name (no .bat)"
+   end
   end
   
+  def self.clear_caches!
+    # I only know about one type...
+   bin_dir = Gem.ruby.split('/')[0..-2].join('/') # strip off ruby.exe
+   Dir[bin_dir = '/*_bin_location'].each{|file|
+     puts 'faster_gem_scripts clearing cached file ' + file
+     File.delete file   
+   }
+    
+  end
+
 end
